make default argument of update-component optional
the default argument is only useful in the case that we query on a
different set of components and (some of) the entities returned are
missing the component we want to update. In my current usage that has
not happened. With this commit on can leave the last argument out and
it will be substituted by #f